Separating out ResourceClass.
